refactor: import public symbols from the package barrel in tests
